Cuda: Use pinned memory for host forces/torques
in preparation for asynchroneous device to host copy.

The host forces/.torques cannot be made local variables, because
the allocation of pinned memory takes longer than the copy operation.

Registrtion of an exit handler in Python is needed, because otherwise
it cannot be guaranteed that the vectors are de-allocated before the
Cuda driver shuts down.